pyproject.toml
[project]
name = "flowautomate-task"
version = "0.1.0"
description = "FastAPI backend scaffold for websocket chat application"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"alembic>=1.18.4",
	"asyncpg>=0.31.0",
	"fastapi>=0.116.1",
	"google-cloud-storage>=3.10.1",
	"greenlet>=3.3.2",
	"gsutil>=5.36",
	"ngrok>=1.7.0",
	"psycopg2-binary>=2.9.11",
	"pydantic>=2.11.7",
	"pydantic-settings>=2.13.1",
	"python-dotenv>=1.2.2",
	"redis>=7.4.0",
	"sqlalchemy>=2.0.49",
	"streamlit>=1.49.1",
	"uvicorn>=0.35.0",
	"websockets>=16.0",
]

[project.scripts]
chat-backend = "main:main"
